leaf will not throw error when finding worker, because we now send empty status updates
