+++ hyperFakeWorker/worker/main.py
@@ -1,5 +1,5 @@
 #!/bin/python
-from .grpc import add_proto_definitions
+from ._grpc import add_proto_definitions
 add_proto_definitions()
 
 import click
@@ -10,18 +10,20 @@ def main():
 
 @main.command()
 def server():
-    from .grpc.server import serve
+    from .server.server import serve
 
     serve()
 
 @main.command()
 def client():
     import grpc
-    from .grpc.controller.controller_pb2_grpc import ControllerStub
+    from ._grpc.controller.controller_pb2_grpc import ControllerStub
+    from ._grpc.controller.controller_pb2 import StatusRequest
     channel = grpc.insecure_channel("localhost:50051")
     stub = ControllerStub(channel)
 
-    status_future = stub.Status()
+    status_future = stub.Status(StatusRequest(nodeID="0"))
+    status_future.result()
 
 
 if __name__ == "__main__":
